Move loadSnapshot to StateStoreArrowFileStore
